Removed CKShapeView
There is a serious bug in `CKShapeView` that hasn't been addressed for some time. It's a simple library so instead we can just animate the shape layer manually.

As an added bonus, we animate a plain `TNKActivityIndicatorView` a little more cleanly now. The thickness of the control was changed to be a single pixel instead of point.